completed search; initial full solution
